refactor(spf): unify track switching into setupTrackSwitching
Merges the two near-identical helpers (setupQualitySwitching for video,
setupAudioTrackSwitching for audio) into one variant-agnostic
setupTrackSwitching. The two variants — switchVideoTrack (bandwidth-driven
ABR) and switchAudioTrack (pin-to-current, ABR-ready shape) — now share
one body and differ only in their per-variant selectOptimal.

- New file packages/spf/src/playback/behaviors/track-switching.ts hosts
  the unified helper + both variant exports + audio-side helpers
  (defaultAudioPicker, selectAudioCurrent). The helper's state map keeps
  bandwidthState so a future bandwidth-driven audio selectOptimal can drop
  in without re-shaping the helper or its state map.
- selectQuality (media/abr/quality-selection.ts) signature changes from
  (tracks, bandwidth, opts) to (tracks, ctx) where ctx packs bandwidth,
  safetyMargin, upgradeMargin, and currentTrack — matches the helper's
  SelectionCtx contract so selectQuality can be passed as a variant's
  selectOptimal directly. Internal caller in media/primitives/select-tracks.ts
  updated.
- switchVideoQuality renamed to switchVideoTrack for naming symmetry with
  switchAudioTrack. Future switchTextTrack slots into the same pattern.
- Deletes quality-switching.ts, switch-audio-track.ts, and their test
  files. Test cases merged into tests/track-switching.test.ts.
- all.ts re-exports the new symbols; engine compositions (engine.ts,
  engine-audio-only.ts) updated. Doc references across calculate-
  presentation-duration, dom/setup-buffer-actors, dom/sync-text-tracks,
  select-tracks, and create-composition updated.

Bundle: 44.64 → 44.03 KB min (-0.61); 12.64 → 12.41 KB gzip (-0.23).
Tests: 895/905 passing (10 pre-existing skipped); typecheck clean.
